requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
faiss-cpu==1.13.0
fastapi==0.123.0
filelock==3.20.0
fonttools==4.52.1
fsspec==2025.10.0
h11==0.16.0
h5py==3.15.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.5
MarkupSafe==3.0.3
matplotlib==3.9.0
mpmath==1.3.0
mysql-connector-python==9.5.0
networkx==3.6
numpy==1.26.4
packaging==24.0
pandas==2.3.3
pillow==10.3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
six==1.16.0
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
